arch(engine): split runner.go SRP violation into sibling files
internal/engine/runner.go reached 1 290 lines mixing seven concerns.
Split the Layer-0 gate, config-resolution cache, and logging helpers
into runner_layer0.go, runner_cache.go, and runner_log.go respectively.
Primary file is now 931 lines. Pure file reorganisation — no logic
changed. Verified by go build ./... and go test ./... (all green).

Files plans 2606211908, 2606211909, 2606211910 for the remaining
three tax items from the 2026-06-21 audit (layer0.go split,
server.go split, workspace.go exemptions). Updates audit log and
marks plan/2606211907 as complete.
